rename sample to show labels on layer in 3d
