--- scene/display-3d-labels-in-scene/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'display-3d-labels-in-scene'
