--- scene/display-3d-labels-in-scene/src/main/java/com/esri/samples/display_3d_labels_in_scene/Display3dLabelsInSceneLauncher.java
@@ -1,11 +1,0 @@
-package com.esri.samples.display_3d_labels_in_scene;
-
-/**
- * Wrapper required for launching a JavaFX 11 app through Gradle or from a jar.
- */
-
-public class Display3dLabelsInSceneLauncher {
-
-  public static void main(String[] args) { Display3dLabelsInSceneSample.main(args); }
-
-}
