+++ scene/show-labels-on-layer-in-3d/gradle/wrapper/gradle-wrapper.properties
