+++ scene/show-labels-on-layer-in-3d/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'show-labels-on-layer-in-3d'
