+++ scene/show-labels-on-layer-in-3d/src/main/java/com/esri/samples/show_labels_on_layer_in_3d/ShowLabelsOnLayerIn3dLauncher.java
@@ -1,0 +1,11 @@
+package com.esri.samples.show_labels_on_layer_in_3d;
+
+/**
+ * Wrapper required for launching a JavaFX 11 app through Gradle or from a jar.
+ */
+
+public class ShowLabelsOnLayerIn3dLauncher {
+
+  public static void main(String[] args) { ShowLabelsOnLayerIn3dSample.main(args); }
+
+}
